chore: refactoring mock data for testing
